+++ src/command_modules/azure-cli-vm/HISTORY.rst
@@ -5,6 +5,7 @@ Release History
 2.2.15
 ++++++
 * vm list-skus: Allow use of `--all` in place of `--all true`
+* Add `vmss run-command [invoke / list / show]`
 
 2.2.14
 ++++++